updated todo enpoints with authentication/authorization

+++ routers/todos.py
@@ -4,6 +4,7 @@ from sqlalchemy.orm import Session
 from fastapi import APIRouter, Depends, HTTPException, Path, status
 from models import Todos
 from db import SessionLocal
+from .auth import get_current_user
 
 from fastapi import APIRouter
 
@@ -18,6 +19,7 @@ def get_db():
         db.close()
 
 db_dependancy = Annotated[Session, Depends(get_db)]
+user_dependancy = Annotated[dict, Depends(get_current_user)]
 
 class TodoRequest(BaseModel):
     title:str = Field(min_length=3)
@@ -27,12 +29,19 @@ class TodoRequest(BaseModel):
 
 
 @router.get("/", status_code=status.HTTP_200_OK)
-async def read_all(db: db_dependancy):
-    return db.query(Todos).all()
+async def read_all(user: user_dependancy, db: db_dependancy):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication Failed")
+    return db.query(Todos).filter(Todos.owner_id == user.get('id')).all()
+
+    
 
 @router.get("/todo/{todo_id}", status_code=status.HTTP_200_OK)
-async def read_todo(db: db_dependancy, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def read_todo(user: user_dependancy, db: db_dependancy, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication Failed")
+    
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
 
     if todo_model is not None:
 
@@ -40,15 +49,22 @@ async def read_todo(db: db_dependancy, todo_id: int = Path(gt=0)):
     raise HTTPException(status_code=404, detail="Todo not found")
 
 @router.post("/todo", status_code=status.HTTP_201_CREATED)
-async def create_todo(db: db_dependancy, todo_request: TodoRequest):
-    todo_model = Todos(**todo_request.model_dump())
+async def create_todo(user: user_dependancy, db: db_dependancy, todo_request: TodoRequest):
+
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication Failed")
+    
+    todo_model = Todos(**todo_request.model_dump(), owner_id = user.get('id'))
 
     db.add(todo_model)
     db.commit()
 
 @router.put("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db: db_dependancy, todo_request:TodoRequest, todo_id:int = Path(gt=0), ):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def update_todo(user: user_dependancy, db: db_dependancy, todo_request:TodoRequest, todo_id:int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication Failed")
+    
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
 
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo not found")
@@ -62,12 +78,15 @@ async def update_todo(db: db_dependancy, todo_request:TodoRequest, todo_id:int =
     db.commit()
 
 @router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependancy, todo_id:int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def delete_todo(user: user_dependancy, db: db_dependancy, todo_id:int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication Failed")
+    
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
     
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo not found")
     
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).delete()
 
     db.commit()
